requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_tflow_select=2.1.0=gpu
absl-py=0.11.0=py36h06a4308_0
affine=2.3.0=pypi_0
astor=0.8.1=py36_0
attrs=20.3.0=pypi_0
blas=1.0=mkl
c-ares=1.17.1=h27cfd23_0
ca-certificates=2021.1.19=h06a4308_0
certifi=2020.12.5=py36h06a4308_0
click=7.1.2=pypi_0
click-plugins=1.1.1=pypi_0
cligj=0.7.1=pypi_0
cudatoolkit=9.0=h13b8566_0
cudnn=7.6.5=cuda9.0_0
cupti=9.0.176=0
cycler=0.10.0=pypi_0
decorator=4.4.2=pypi_0
freetype=2.10.4=h5ab3b9f_0
gast=0.4.0=py_0
grpcio=1.31.0=py36hf8bcb03_0
h5py=2.10.0=py36hd6299e0_1
hdf5=1.10.6=hb1b8bf9_0
imageio=2.9.0=pypi_0
importlib-metadata=2.0.0=py_1
intel-openmp=2020.2=254
jpeg=9b=h024ee3a_2
kiwisolver=1.3.1=pypi_0
lcms2=2.11=h396b838_0
ld_impl_linux-64=2.33.1=h53a641e_7
libedit=3.1.20191231=h14c3975_1
libffi=3.3=he6710b0_2
libgcc-ng=9.1.0=hdf63c60_0
libgfortran-ng=7.3.0=hdf63c60_0
libpng=1.6.37=hbc83047_0
libprotobuf=3.13.0.1=hd408876_0
libstdcxx-ng=9.1.0=hdf63c60_0
libtiff=4.1.0=h2733197_1
lz4-c=1.9.3=h2531618_0
markdown=3.3.3=py36h06a4308_0
matplotlib=3.1.3=pypi_0
mkl=2020.2=256
mkl-service=2.3.0=py36he8ac12f_0
mkl_fft=1.2.0=py36h23d657b_0
mkl_random=1.1.1=py36h0573a6f_0
ncurses=6.2=he6710b0_1
netron=4.7.8=pypi_0
networkx=2.5=pypi_0
numpy=1.16.0=pypi_0
numpy-base=1.19.2=py36hfa32c7d_0
olefile=0.46=py36_0
opencv-python=4.4.0.46=pypi_0
openssl=1.1.1i=h27cfd23_0
pillow=8.1.0=py36he98fc37_0
pip=20.3=py36h06a4308_0
protobuf=3.13.0.1=py36he6710b0_1
pyparsing=2.4.7=pypi_0
python=3.6.12=hcff3b4d_2
python-dateutil=2.8.1=pypi_0
pywavelets=1.1.1=pypi_0
rasterio=1.2.0=pypi_0
readline=8.0=h7b6447c_0
rio-color=1.0.2=pypi_0
rio-mucho=1.0.0=pypi_0
scikit-image=0.17.2=pypi_0
scipy=1.1.0=pypi_0
setuptools=39.1.0=pypi_0
six=1.15.0=py36h06a4308_0
snuggs=1.4.7=pypi_0
sqlite=3.33.0=h62c20be_0
tensorboard=1.9.0=py36hf484d3e_0
tensorflow=1.9.0=gpu_py36h02c5d5e_1
tensorflow-base=1.9.0=gpu_py36h6ecc378_0
tensorflow-gpu=1.9.0=pypi_0
termcolor=1.1.0=py36_1
tifffile=2020.9.3=pypi_0
tk=8.6.10=hbc83047_0
tqdm=4.54.1=pypi_0
werkzeug=1.0.1=py_0
wheel=0.36.0=pyhd3eb1b0_0
xz=5.2.5=h7b6447c_0
zipp=3.4.0=pyhd3eb1b0_0
zlib=1.2.11=h7b6447c_3
zstd=1.4.5=h9ceee32_0
